added makefile, removed format/pointarith warnings when compiling cmp.cpp

+++ src/include/cmp.cpp
@@ -12,7 +12,7 @@
 #include <stdio.h>
 
 CatMemoryPool::CatMemoryPool(void* base, size_t size) : base(base), size(size) {
-	printf("creating memory pool with size 0x%08x at 0x%08x\n", size, base);
+	printf("creating memory pool with size 0x%08x at 0x%08x\n", (unsigned)size, (unsigned)base);
 }
 
 
@@ -57,10 +57,10 @@ CatMemoryPool::pool_block_s* CatMemoryPool::FindBlock(size_t size) {
 void* CatMemoryPool::alloc(size_t size) {
 	pool_block_s* block = FindBlock(size);
 	if (block == (pool_block_s*)-1) {
-		printf("not enough memory to allocate block of size 0x%08x\n", size);
+		printf("not enough memory to allocate block of size 0x%08x\n", (unsigned)size);
 		return (void*)0;
 	}
-	printf("allocating block of size 0x%08x at 0x%08x\n", size, pool_pointer<void>(block));
+	printf("allocating block of size 0x%08x at 0x%08x\n", (unsigned)size, (unsigned)pool_pointer<void>(block));
 	ChipBlock(block, size);
 	block->free = false;
 	return (void*)((uintptr_t)(block) + sizeof(pool_block_s));
@@ -68,7 +68,7 @@ void* CatMemoryPool::alloc(size_t size) {
 
 void CatMemoryPool::free(void* object) {
 	pool_block_s* block = (pool_block_s*)((uintptr_t)object - sizeof(pool_block_s));
-	printf("freeing block of size 0x%08x at 0x%08x\n", block->size, pool_pointer<void>(block));
+	printf("freeing block of size 0x%08x at 0x%08x\n", block->size, (unsigned)pool_pointer<void>(block));
 	block->free = true;
 	MendBlock(block);
 }
@@ -82,12 +82,12 @@ void CatMemoryPool::ChipBlock(pool_block_s* block, size_t size) {
 		new_block.next = block->next;
 		new_block.free = 1;
 		new_block.size = old_size - (size + sizeof(pool_block_s));
-		void* p_new_block = (void*)(pool_pointer<void>(block) + sizeof(pool_block_s) + block->size);
+		void* p_new_block = (void*)((unsigned)pool_pointer<void>(block) + sizeof(pool_block_s) + block->size);
 		if (block->next != (void*)-1) {
 					real_pointer<pool_block_s>(block->next)->prev = p_new_block;
 		}
 		block->next = p_new_block;
-		printf("chipping block at 0x%08x with old size 0x%08x, made new block at 0x%08x with size 0x%08x\n", pool_pointer<void>(block), old_size, p_new_block, new_block.size);
+		printf("chipping block at 0x%08x with old size 0x%08x, made new block at 0x%08x with size 0x%08x\n", (unsigned)pool_pointer<void>(block), old_size, (unsigned)p_new_block, new_block.size);
 		memcpy(real_pointer<void>(p_new_block), &new_block, sizeof(pool_block_s));
 	}
 }
@@ -95,7 +95,7 @@ void CatMemoryPool::ChipBlock(pool_block_s* block, size_t size) {
 void CatMemoryPool::print() {
 	pool_block_s* current = (pool_block_s*)base;
 	while (true) {
-		printf("] memory block at 0x%08x (0x%08x) size 0x%08x next 0x%08x prev 0x%08x free? %i. \n", pool_pointer<void>(current), current, current->size, current->next, current->prev, current->free);
+		printf("] memory block at 0x%08x (0x%08x) size 0x%08x next 0x%08x prev 0x%08x free? %i. \n", (unsigned)pool_pointer<void>(current), (unsigned)current, current->size, (unsigned)current->next, (unsigned)current->prev, current->free);
 		if (current->next == (void*)-1) break;
 		current = real_pointer<pool_block_s>(current->next);
 	}
@@ -113,7 +113,7 @@ void CatMemoryPool::MendBlock(pool_block_s* block) {
 		pool_block_s* cur_next = real_pointer<pool_block_s>(block->next);
 		while (cur_next->free) {
 			block->size += sizeof(pool_block_s) + cur_next->size;
-			printf("mended block at 0x%08x new size 0x%08x\n", pool_pointer<void>(block), block->size);
+			printf("mended block at 0x%08x new size 0x%08x\n", (unsigned)pool_pointer<void>(block), block->size);
 			DeleteBlock(cur_next);
 			if (block->next != (void*)-1) {
 				cur_next = real_pointer<pool_block_s>(block->next);
@@ -123,7 +123,7 @@ void CatMemoryPool::MendBlock(pool_block_s* block) {
 }
 
 void CatMemoryPool::DeleteBlock(pool_block_s* block) {
-	printf("deleted block at 0x%08x size 0x%08x\n", pool_pointer<void>(block), block->size);
+	printf("deleted block at 0x%08x size 0x%08x\n", (unsigned)pool_pointer<void>(block), block->size);
 	if (block->next != (void*)-1) real_pointer<pool_block_s>(block->next)->prev = block->prev;
 	if (block->prev != (void*)-1) real_pointer<pool_block_s>(block->prev)->next = block->next;
 }
